requirements.txt
pyautogui==0.9.54
opencv-python==4.12.0.88
pillow==12.1.1
paddlepaddle==3.2.2
paddleocr==3.2.0
keyboard==0.13.5
chardet==5.2.0
